+++ x/leverage/types/errors.go
@@ -7,38 +7,44 @@ import (
 )
 
 var (
-	ErrInvalidAsset            = sdkerrors.Register(ModuleName, 1100, "invalid asset")
-	ErrInsufficientBalance     = sdkerrors.Register(ModuleName, 1101, "insufficient balance")
-	ErrUndercollaterized       = sdkerrors.Register(ModuleName, 1102, "borrow positions are undercollaterized")
-	ErrLendingPoolInsufficient = sdkerrors.Register(ModuleName, 1103, "lending pool insufficient")
-	ErrInvalidRepayment        = sdkerrors.Register(ModuleName, 1104, "invalid repayment")
-	ErrInvalidAddress          = sdkerrors.Register(ModuleName, 1105, "invalid address")
-	ErrNegativeTotalBorrowed   = sdkerrors.Register(ModuleName, 1106, "total borrowed was negative")
-	ErrInvalidUtilization      = sdkerrors.Register(ModuleName, 1107, "invalid token utilization")
-	ErrLiquidationIneligible   = sdkerrors.Register(ModuleName, 1108, "borrower not eligible for liquidation")
-	ErrBadValue                = sdkerrors.Register(ModuleName, 1109, "bad USD value")
-	ErrLiquidatorBalanceZero   = sdkerrors.Register(ModuleName, 1110, "liquidator base asset balance is zero")
-	ErrNegativeTimeElapsed     = sdkerrors.Register(ModuleName, 1111, "negative time elapsed since last interest time")
-	ErrInvalidOraclePrice      = sdkerrors.Register(ModuleName, 1112, "invalid oracle price")
-	ErrNegativeAPY             = sdkerrors.Register(ModuleName, 1113, "negative APY")
-	ErrInvalidExchangeRate     = sdkerrors.Register(ModuleName, 1114, "exchange rate less than one")
-	ErrInconsistentTotalBorrow = sdkerrors.Register(ModuleName, 1115, "total adjusted borrow inconsistency")
-	ErrInvalidInteresrScalar   = sdkerrors.Register(ModuleName, 1116, "interest scalar less than one")
-	ErrEmptyAddress            = sdkerrors.Register(ModuleName, 1117, "empty address")
-	ErrLiquidationRewardRatio  = sdkerrors.Register(ModuleName, 1118, "requested liquidation reward not met")
-	ErrSupplyNotAllowed        = sdkerrors.Register(ModuleName, 1119, "supplying of asset disabled")
-	ErrBorrowNotAllowed        = sdkerrors.Register(ModuleName, 1120, "borrowing of asset disabled")
-	ErrBlacklisted             = sdkerrors.Register(ModuleName, 1121, "base denom blacklisted")
-	ErrCollateralWeightZero    = sdkerrors.Register(ModuleName, 1122, "token collateral weight is zero")
-	ErrLiquidationInvalid      = sdkerrors.Register(ModuleName, 1123, "liquidation invalid")
-	ErrMaxSupplyUtilization    = sdkerrors.Register(ModuleName, 1124, "market would exceed MaxSupplyUtilization")
-	ErrMinCollateralLiquidity  = sdkerrors.Register(ModuleName, 1125, "market would fall below MinCollateralLiquidity")
-	ErrMaxCollateralShare      = sdkerrors.Register(
-		ModuleName,
-		1126,
-		"market total collateral would exceed MaxCollateralShare",
-	)
-	ErrNotUToken            = sdkerrors.Register(ModuleName, 1127, "denom should be a uToken")
-	ErrUToken               = sdkerrors.Register(ModuleName, 1128, "denom should not be a uToken")
-	ErrExcessiveTimeElapsed = sdkerrors.Register(ModuleName, 1130, "excessive time elapsed since last interest time")
+	// 1XX = General Validation
+	ErrEmptyAddress = sdkerrors.Register(ModuleName, 100, "empty address")
+	ErrNilAsset     = sdkerrors.Register(ModuleName, 101, "nil asset")
+
+	// 2XX = Token Registry
+	ErrNotRegisteredToken   = sdkerrors.Register(ModuleName, 200, "not a registered Token")
+	ErrUToken               = sdkerrors.Register(ModuleName, 201, "denom should not be a uToken")
+	ErrNotUToken            = sdkerrors.Register(ModuleName, 202, "denom should be a uToken")
+	ErrSupplyNotAllowed     = sdkerrors.Register(ModuleName, 203, "supplying of Token disabled")
+	ErrBorrowNotAllowed     = sdkerrors.Register(ModuleName, 204, "borrowing of Token disabled")
+	ErrBlacklisted          = sdkerrors.Register(ModuleName, 205, "blacklisted Token")
+	ErrCollateralWeightZero = sdkerrors.Register(ModuleName, 206, "collateral weight of Token is zero")
+
+	// 3XX = User Positions
+	ErrInsufficientBalance    = sdkerrors.Register(ModuleName, 300, "insufficient balance")
+	ErrInsufficientCollateral = sdkerrors.Register(ModuleName, 301, "insufficient collateral")
+	ErrDenomNotBorrowed       = sdkerrors.Register(ModuleName, 302, "denom not borrowed")
+	ErrLiquidationRepayZero   = sdkerrors.Register(ModuleName, 303, "liquidation would repay zero tokens")
+
+	// 4XX = Price Sensitive
+	ErrBadValue              = sdkerrors.Register(ModuleName, 400, "bad USD value")
+	ErrInvalidOraclePrice    = sdkerrors.Register(ModuleName, 401, "invalid oracle price")
+	ErrUndercollaterized     = sdkerrors.Register(ModuleName, 402, "borrow positions are undercollaterized")
+	ErrLiquidationIneligible = sdkerrors.Register(ModuleName, 403, "borrower not eligible for liquidation")
+
+	// 5XX = Market Conditions
+	ErrLendingPoolInsufficient = sdkerrors.Register(ModuleName, 500, "lending pool insufficient")
+	ErrMaxSupplyUtilization    = sdkerrors.Register(ModuleName, 501, "market would exceed MaxSupplyUtilization")
+	ErrMinCollateralLiquidity  = sdkerrors.Register(ModuleName, 502, "market would fall below MinCollateralLiquidity")
+	ErrMaxCollateralShare      = sdkerrors.Register(ModuleName, 503, "market would exceed MaxCollateralShare")
+
+	// 6XX = Internal Failsafes
+	ErrInvalidUtilization      = sdkerrors.Register(ModuleName, 600, "invalid token utilization")
+	ErrNegativeTotalBorrowed   = sdkerrors.Register(ModuleName, 601, "total borrowed was negative")
+	ErrNegativeAPY             = sdkerrors.Register(ModuleName, 602, "negative APY")
+	ErrNegativeTimeElapsed     = sdkerrors.Register(ModuleName, 603, "negative time elapsed since last interest time")
+	ErrInvalidExchangeRate     = sdkerrors.Register(ModuleName, 604, "exchange rate less than one")
+	ErrInconsistentTotalBorrow = sdkerrors.Register(ModuleName, 605, "total adjusted borrow inconsistency")
+	ErrInvalidInteresrScalar   = sdkerrors.Register(ModuleName, 606, "interest scalar less than one")
+	ErrExcessiveTimeElapsed    = sdkerrors.Register(ModuleName, 607, "excessive time elapsed since last interest time")
 )